+++ bot_client.cpp
@@ -822,12 +822,12 @@ void BotClient_Valve_ScreenFade(void *p, int bot_index) {
     } else if (state == 6) {
         state = 0;
 
-        float length = (duration + hold_time) / 4096.0f;
+        float length = (static_cast<float>duration + hold_time) / 4096.0f;
         int iDevide = bots[bot_index].bot_skill;
         if (iDevide < 1)
             iDevide = 1;
 
-        length -= ((10 / iDevide) * 0.5f);
+        length -= ((static_cast<float>(10) / iDevide) * 0.5f);
 
         bots[bot_index].fBlindedTime = gpGlobals->time + length;
